Biryani_classes
According to a recent survey, Biryani is the most ordered food. The chef wants to learn how to make world-class Biryani from a MasterChef. The chef will be required to attend MasterChef's classes for X weeks, and the cost of classes per week is Y coins. What is the total amount of money that Chef will have to pay?

Input Format
The first line of input will contain an integer 𝑇 the number of test cases. The description of 𝑇 test cases follows. The first and only line of each test case contains two space-separated integers X and Y, as described in the problem statement.

Output Format
For each test case, output on a new line the total amount of money that the Chef will have to pay.
Input                      Output
4                
1 10                        10
1 15                        15
2 10                        20
2 15                        30