You are given an assignment from your institute.
Each assignment has a difficulty value between 1 and 100 (both inclusive).
AI can solve an assignment if and only if the difficulty of the assignment is less than or equal to 60.

Your current assignment has difficulty
X. Is AI capable of solving your assignment? Print "YES" if it is, and "NO" if it isn't (without the quotes).